// src/shared/api/generated/model/productSubscriptionExecuteRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */
import type { AgreedTermsItem } from './agreedTermsItem';

export interface ProductSubscriptionExecuteRequest {
  productId: number;
  subscriptionAmount: number;
  termMonths: number;
  withdrawalAccountId: number;
  /** @pattern ^[0-9]{4}$ */
  newAccountPassword: string;
  /** @pattern ^[0-9]{4}$ */
  newAccountPasswordConfirm: string;
  /** @minLength 1 */
  accountPasswordAuthToken: string;
  /** @minLength 1 */
  otpAuthToken: string;
  agreedTerms: AgreedTermsItem[];
}