// src/shared/api/generated/model/productSubscriptionValidationRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */
import type { AgreedTermsItem } from './agreedTermsItem';

export interface ProductSubscriptionValidationRequest {
  productId: number;
  subscriptionAmount: number;
  termMonths: number;
  withdrawalAccountId: number;
  agreedTerms: AgreedTermsItem[];
  satisfiedConditionCodes?: string[];
}