// src/shared/api/generated/model/productSubscriptionValidationResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */
import type { ProductSubscriptionValidationResponseProductGroup } from './productSubscriptionValidationResponseProductGroup';
import type { ViolationItem } from './violationItem';

export interface ProductSubscriptionValidationResponse {
  valid?: boolean;
  violations?: ViolationItem[];
  productGroup?: ProductSubscriptionValidationResponseProductGroup;
  baseRate?: number;
  preferentialRate?: number;
  appliedRate?: number;
  maturityDate?: string;
  expectedPrincipal?: number;
  expectedInterest?: number;
  expectedMaturityAmount?: number;
  withdrawalAccountNumber?: string;
  withdrawalAccountBalance?: number;
}